src/pdf_extraction_app/requirements.txt
SQLAlchemy==2.0.40
pgvector==0.4.0
psycopg2-binary==2.9.10
nltk==3.9.1
confluent-kafka==2.9.0
PyMuPDF==1.25.2
google-generativeai==0.8.4
sentence-transformers==3.4.1